bg(ownerId):fix get account error create a new class to get owner account by Id imported the class in account models [#165645069]
