+++ package.json
@@ -9,7 +9,7 @@
     "devTest": "cross-env TEST_ENV=true babel-node server/create_tables.js && cross-env TEST_ENV=true nyc --require @babel/register  mocha server/unitTest --timeout 10000 --exit",
     "singleTest": "cross-env TEST_ENV=true babel-node server/create_tables.js && cross-env TEST_ENV=true nyc --require @babel/register  mocha --timeout 10000 --exit",
     "create": "node ./server/database/database.main.js",
-    "start": "babel ./server -d build && node build/app.js",
+    "start":"rimraf dist/ && babel  ./server -d dist && node dist/app.js",
     "start-node": "node ./build/app.js",
     "dev": "nodemon --exec babel-node server/app.js",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
