+++ server/database/models/accounts.js
@@ -1,5 +1,72 @@
 import databaseController from "../database";
 
+
+exports.debitUser = async (data, callbk) => {
+  try {
+    const {
+      amount, accountNumber, cashierId,
+    } = data;
+
+    const account = await databaseController.findAccountByAccountNumber(accountNumber);
+
+    if (!cashierId) { callbk({ message: "Account Not Found", code: 404 }, null); return; }
+    if (!account) { callbk({ message: "Account Not Found", code: 404 }, null); return; }
+    if (account.balance < data.amount) { callbk({ message: "Insufficient Funds", code: 400 }, null); return; }
+    if (account.status === "dormant") { callbk({ message: "Account Dormant", code: 400 }, null); return; }
+
+    const type = "debit";
+    const depositor = "N/A";
+    const oldBalance = account.balance;
+    const newBalance = oldBalance - Number(data.amount);
+    const updatedAccount = await databaseController.updateAccountbalance({ accountNumber, newBalance });
+    const transactionDetail = await databaseController.saveTransaction({
+      accountNumber,
+      amount: Number(amount),
+      cashierId,
+      depositor,
+      type,
+      oldBalance,
+      newBalance: updatedAccount.balance,
+    });
+    delete transactionDetail.depositor;
+    callbk(null, transactionDetail);
+  } catch (err) {
+    callbk({ message: err.message.replace(/[^\w|\s]/g, ""), code: 400 }, null);
+  }
+};
+
+exports.creditUser = async (data, callbk) => {
+  const {
+    amount, accountNumber, depositor, cashierId,
+  } = data;
+
+  try {
+    const account = await databaseController.findAccountByAccountNumber(accountNumber);
+
+    if (!cashierId) { callbk({ message: "Account Not Found", code: 404 }, null); return; }
+    if (!account) { callbk({ message: "Account Not Found", code: 404 }, null); return; }
+    if (account.status === "dormant") { callbk({ message: "Account Dormant", code: 400 }, null); return; }
+
+    const type = "credit";
+    const oldBalance = account.balance;
+    const newBalance = oldBalance + Number(data.amount);
+    const updatedAccount = await databaseController.updateAccountbalance({ accountNumber, newBalance });
+    const transactionDetail = await databaseController.saveTransaction({
+      accountNumber,
+      amount: Number(amount),
+      cashierId,
+      depositor,
+      type,
+      oldBalance,
+      newBalance: updatedAccount.balance,
+    });
+    callbk(null, transactionDetail);
+  } catch (err) {
+    callbk({ message: err.message.replace(/[^\w|\s]/g, ""), code: 400 }, null);
+  }
+};
+
+
 exports.DeleteAccount = async (data, callbck) => {
   try {
     const account = await databaseController.findAccountByAccountNumber(data);
@@ -13,6 +80,7 @@ exports.DeleteAccount = async (data, callbck) => {
   }
 };
 
+
 exports.getAcountByAccountNumber = async (data, callbk) => {
   try {
     const account = await databaseController.findAccountByAccountNumber(data);
@@ -23,12 +91,14 @@ exports.getAcountByAccountNumber = async (data, callbk) => {
   }
 };
 
+
 exports.getAllAccountsByOwnerid = async (data, callbk) => {
   try {
     const { currentUser } = data.req;
     const result = await databaseController.findUserById(currentUser.id);
     const accounts = await databaseController.findAccountByOwnerid(data.userId);
     const account = accounts.find(value => value.ownerid === data.userId);
+    if (!account) { callbk({ message: "No account found" }, null); return; }
     if (result.isadmin || result.id === account.ownerid) {
       callbk(null, accounts);
       return;
@@ -55,7 +125,6 @@ exports.getAcountByEmail = async (data, callbk) => {
     const { currentUser } = data.req;
     const result = await databaseController.findUserById(currentUser.id);
     const accounts = await databaseController.findAccountByEmail(data.useEmail);
-    // const accounts = await databaseController.findAccountByOwnerid(data.userId);
     const account = accounts.find(value => value.ownerid === data.userId);
     if (result.isadmin || result.id === account.ownerid) {
       callbk(null, accounts);
@@ -63,8 +132,23 @@ exports.getAcountByEmail = async (data, callbk) => {
     }
     callbk({ message: "Forbidden", code: 403 }, null);
     return;
-    // if (!accounts) { callbk({ message: "Account not found" }, null); return; }
-    // callbk(null, account);
+  } catch (err) {
+    callbk({ message: err.message.replace(/[^\w|\s]/g, "") }, null);
+  }
+};
+
+exports.toggleAccountStatus = async (data, callbk) => {
+  const account = await databaseController.findAccountByAccountNumber(data);
+
+  try {
+    if (!account) { callbk({ message: "Account not found", code: 404 }, null); return; }
+
+    const status = account.status === "active" ? "dormant" : "active";
+    const { accountnumber } = account;
+
+
+    const update = await databaseController.updateAccountStatus({ status, accountnumber });
+    callbk(null, { update });
   } catch (err) {
     callbk({ message: err.message.replace(/[^\w|\s]/g, "") }, null);
   }
