+++ server/middleware/validation.js
@@ -1,6 +1,5 @@
 import Joi from "joi";
 
-
 exports.login = (req, res, next) => {
   const SchemaLogin = Joi.object().keys({
     email: Joi.string().email({ minDomainAtoms: 2 }).required(),
@@ -46,6 +45,36 @@ exports.adminCreate = (req, res, next) => {
     email: Joi.string().email({ minDomainAtoms: 2 }).required(),
     phonenumber: Joi.string().regex(/^\d{3}-\d{4}-\d{6}$/).required(),
     type: Joi.string().alphanum().min(3).max(8),
+  });
+  const response = Joi.validate(req.body, SchemaLogin);
+  if (response.error) {
+    return res.status(400).json({
+      status: 400,
+      error: response.error.details[0].message.replace(/[^\w|\s]/g, ""),
+    });
+  }
+  next();
+};
+
+exports.credit = (req, res, next) => {
+  const SchemaLogin = Joi.object().keys({
+    depositor: Joi.string().min(7).max(30)
+      .required(),
+    amount: Joi.number().positive().required(),
+    type: Joi.string().alphanum().min(3).max(8),
+  });
+  const response = Joi.validate(req.body, SchemaLogin);
+  if (response.error) {
+    return res.status(400).json({
+      status: 400,
+      error: response.error.details[0].message.replace(/[^\w|\s]/g, ""),
+    });
+  }
+  next();
+};
+exports.debit = (req, res, next) => {
+  const SchemaLogin = Joi.object().keys({
+    amount: Joi.number().positive().required(),
   });
   const response = Joi.validate(req.body, SchemaLogin);
   if (response.error) {
